Move html_assert tests to named file
